프로젝트3팀-미니언즈_Django프로젝트소스코드_V1.0/requirements.txt
absl-py==2.1.0
albucore==0.0.12
albumentations==1.4.12
amqp==5.2.0
annotated-types==0.7.0
asgiref==3.8.1
asttokens==2.4.1
attrs==24.2.0
beautifulsoup4==4.12.3
billiard==4.2.0
celery==5.4.0
certifi==2024.7.4
cffi==1.17.0
channels==4.1.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.2.1
cryptography==43.0.0
cycler==0.12.1
Cython==3.0.10
debugpy==1.8.2
decorator==5.1.1
Django==4.2.15
django-bootstrap-v5==1.0.11
django-bootstrap5==24.2
django-phonenumber-field==8.0.0
django-widget-tweaks==1.5.0
dlib==19.24.6
easydict==1.13
eval_type_backport==0.2.0
executing==2.0.1
facenet-pytorch==2.5.3
filelock==3.15.4
flatbuffers==24.3.25
fonttools==4.53.1
fsspec==2024.6.1
humanfriendly==10.0
idna==3.7
imageio==2.34.2
insightface==0.7.3
ipykernel==6.29.5
ipython==8.26.0
jax==0.4.31
jaxlib==0.4.31
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
kombu==5.4.0
lazy_loader==0.4
MarkupSafe==2.1.5
matplotlib==3.9.1
matplotlib-inline==0.1.7
mediapipe==0.10.14
ml-dtypes==0.4.0
mpmath==1.3.0
mysqlclient==2.2.4
nest-asyncio==1.6.0
networkx==3.3
numpy==1.24.4
onnx==1.16.1
onnxruntime==1.18.1
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt-einsum==3.3.0
packaging==24.1
pandas==2.2.2
parso==0.8.4
phonenumbers==8.13.43
pillow==10.4.0
platformdirs==4.2.2
prettytable==3.10.2
prompt_toolkit==3.0.47
protobuf==4.25.4
psutil==6.0.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
PyMySQL==1.1.1
pyparsing==3.1.2
pyreadline3==3.4.1
python-dateutil==2.9.0.post0
pytz==2024.1
pywin32==306
PyYAML==6.0.1
pyzmq==26.0.3
rabbitmq==0.2.0
requests==2.32.3
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.14.0
six==1.16.0
sounddevice==0.5.0
soupsieve==2.5
sqlparse==0.5.1
stack-data==0.6.3
sympy==1.13.1
threadpoolctl==3.5.0
tifffile==2024.7.24
tomli==2.0.1
torch==2.4.0
torchaudio==2.4.0
torchvision==0.19.0
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
